moved event-reminder-selector inside calendar module removed MeetingDetailsModal import from my-meetings module added createdByUser field to Meeting details and meeting create
