+++ eta-company-calendar-frontend/src/app/calendar/modals/event-reminder-selector.component.ts
@@ -1,16 +1,31 @@
-import { Component, Inject, OnInit } from '@angular/core';
+import { Component, Inject } from '@angular/core';
 import { MAT_DIALOG_DATA, MatDialog } from '@angular/material/dialog';
 import { EventInput } from '@fullcalendar/core';
 import { Observable } from 'rxjs';
 import { UserResponse } from '~/app/models/user-response.model';
-import { MeetingCreateComponent } from './../calendar/modals/meeting-create.component';
+import { MeetingCreateComponent } from './meeting-create.component';
 
 @Component({
   selector: 'app-event-reminder-selector',
-  templateUrl: './event-reminder-selector.component.html',
-  styleUrls: ['./event-reminder-selector.component.scss']
+  template: `
+    <div align=center>
+      <h2>{{'eventselector.dialog_title' | translate | uppercase}}</h2>
+      <h4>{{'eventselector.selector_title' | translate | uppercase}}</h4>
+    </div>
+    <a mat-stroked-button (click)=newMeeting() mat-dialog-close>{{'eventselector.new_meeting' | translate}}</a>
+    <br>
+    <a mat-stroked-button >{{'eventselector.new_reminder' | translate}}</a>
+  `,
+  styles: [`
+    a {
+      border: 2px solid;
+      border-color: black !important;
+      width: 200px;
+      margin-bottom: 10%;
+    }
+  `]
 })
-export class EventReminderSelectorComponent implements OnInit {
+export class EventReminderSelectorComponent {
 
   protected loggedInUser: UserResponse;
   protected selectedUser: UserResponse;
@@ -18,9 +33,6 @@ export class EventReminderSelectorComponent implements OnInit {
 
   constructor(private readonly dialog: MatDialog,
               @Inject(MAT_DIALOG_DATA) private readonly data: DialogData) { }
-
-  // tslint:disable-next-line: no-empty
-  public ngOnInit(): void {}
 
   protected newMeeting() {
     this.dialog.open(MeetingCreateComponent, {
