--- eta-company-calendar-frontend/src/app/event-reminder-selector/event-reminder-selector.component.html
@@ -1,9 +1,0 @@
-
-
-<div align=center>
-    <h2>{{'eventselector.dialog_title' | translate | uppercase}}</h2>
-    <h4>{{'eventselector.selector_title' | translate | uppercase}}</h4>
-</div>
-<a mat-stroked-button (click)=newMeeting() mat-dialog-close>{{'eventselector.new_meeting' | translate}}</a>
-<br>
-<a mat-stroked-button >{{'eventselector.new_reminder' | translate}}</a>
