--- eta-company-calendar-frontend/src/app/event-reminder-selector/event-reminder-selector.component.scss
@@ -1,6 +1,0 @@
-a {
-    border: 2px solid;
-    border-color: black !important;
-    width: 200px;
-    margin-bottom: 10%;
-  }
