+++ eta-company-calendar-frontend/src/app/models/meeting-detail.model.ts
@@ -12,6 +12,7 @@ export interface MeetingDetail {
   startingTime: number;
   finishTime: number;
   createdBy: string;
+  createdByUser: number;
   requiredAttendants: string[];
   optionalAttendants: string[];
 }
