+++ eta-company-calendar-frontend/src/app/my-meetings/my-meetings.module.ts
@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
 import { TranslateModule } from '@ngx-translate/core';
-import { MeetingDetailsModal } from '~/app/shared/modals/meeting-details.component.ts';
 import { SharedModule } from '~/app/shared/shared.module';
 import { MyMeetingsRoutingModule } from '../my-meetings/my-meetings-routing.module';
 import { MyMeetingsDescriptionComponent } from './components/my-meetings-description.component';
@@ -9,7 +8,6 @@ import { MeetingService } from './service/meeting.service';
 @NgModule({
   declarations: [
     MyMeetingsDescriptionComponent,
-    MeetingDetailsModal,
   ],
   imports: [
     SharedModule,
